+++ App.js
@@ -4,8 +4,8 @@ import { Button } from 'react-native';
 import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
 
-import Login from './auth/Login';
-import screens from './constants/screens';
+import Login from './src/screens/login/Login';
+import screens from './src/constants/screens';
 
 const Stack = createStackNavigator();
